+++ src/main/java/com/woowacamp/soolsool/core/receipt/domain/converter/ReceiptPriceConverter.java
@@ -1,6 +1,6 @@
-package com.woowacamp.soolsool.core.receipt.converter;
+package com.woowacamp.soolsool.core.receipt.domain.converter;
 
-import com.woowacamp.soolsool.core.receipt.vo.ReceiptPrice;
+import com.woowacamp.soolsool.core.receipt.domain.vo.ReceiptPrice;
 import java.math.BigInteger;
 import javax.persistence.AttributeConverter;
 import javax.persistence.Converter;
