+++ src/main/java/com/woowacamp/soolsool/core/receipt/dto/ReceiptResponse.java
@@ -1,0 +1,10 @@
+package com.woowacamp.soolsool.core.receipt.dto;
+
+import lombok.Getter;
+import lombok.RequiredArgsConstructor;
+
+@Getter
+@RequiredArgsConstructor
+public class ReceiptResponse {
+
+}
