+++ src/test/java/com/woowacamp/soolsool/core/receipt/vo/ReceiptPriceTest.java
@@ -5,6 +5,7 @@ import static org.assertj.core.api.Assertions.assertThatCode;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 import static org.junit.jupiter.api.Assertions.assertAll;
 
+import com.woowacamp.soolsool.core.receipt.domain.vo.ReceiptPrice;
 import com.woowacamp.soolsool.global.exception.SoolSoolException;
 import java.math.BigInteger;
 import org.junit.jupiter.api.DisplayName;
